+++ app/src/main/java/se/kjellstrand/webshooter/ui/navigation/Screen.kt
@@ -6,14 +6,14 @@ sealed class Screen(val route: String) {
     object LandingScreen : Screen("landing")
     object CompetitionsList : Screen("competitions")
     object CompetitionResults : Screen("competition_results/{competitionId}/{resultsType}?competitionName={competitionName}&competitionDate={competitionDate}") {
-        fun createRoute(competitionId: Int, resultsType: String, competitionName: String, competitionDate: String) =
+        fun createRoute(competitionId: Long, resultsType: String, competitionName: String, competitionDate: String) =
             "competition_results/$competitionId/$resultsType?competitionName=${android.net.Uri.encode(competitionName)}&competitionDate=${android.net.Uri.encode(competitionDate)}"
     }
     object Club : Screen("club")
     object MyEntries : Screen("my_entries")
     object Settings : Screen("settings")
     object ShooterResult : Screen("shooter_result/{competitionId}/{shooterId}/{resultsType}") {
-        fun createRoute(competitionId: Int, shooterId: Int, resultsType: String) =
+        fun createRoute(competitionId: Long, shooterId: Long, resultsType: String) =
             "shooter_result/$competitionId/$shooterId/$resultsType"
     }
     object CompetitionSignup : Screen("competition_signup/{competitionId}") {
